+++ library/src/backend/device/device_tridiagonal.cpp
@@ -49,12 +49,12 @@ void linalg::free_tridiagonal_device_data(tridiagonal_descr* descr)
     }
 }
 
-void linalg::device_tridiagonal_analysis(int                  m,
-                                         int                  n,
-                                         const vector<float>& lower_diag,
-                                         const vector<float>& main_diag,
-                                         const vector<float>& upper_diag,
-                                         tridiagonal_descr*   descr)
+void linalg::device_tridiagonal_analysis(int                   m,
+                                         int                   n,
+                                         const vector<double>& lower_diag,
+                                         const vector<double>& main_diag,
+                                         const vector<double>& upper_diag,
+                                         tridiagonal_descr*    descr)
 {
     ROUTINE_TRACE("linalg::device_tridiagonal_analysis");
     if constexpr(is_cuda_available())
@@ -71,11 +71,11 @@ void linalg::device_tridiagonal_analysis(int                  m,
 
 void linalg::device_tridiagonal_solver(int                      m,
                                        int                      n,
-                                       const vector<float>&     lower_diag,
-                                       const vector<float>&     main_diag,
-                                       const vector<float>&     upper_diag,
-                                       const vector<float>&     b,
-                                       vector<float>&           x,
+                                       const vector<double>&    lower_diag,
+                                       const vector<double>&    main_diag,
+                                       const vector<double>&    upper_diag,
+                                       const vector<double>&    b,
+                                       vector<double>&          x,
                                        const tridiagonal_descr* descr)
 {
     ROUTINE_TRACE("linalg::device_tridiagonal_solver");
