+++ library/src/backend/device/device_tridiagonal.h
@@ -34,19 +34,19 @@ namespace linalg
     struct tridiagonal_descr;
 
     void free_tridiagonal_device_data(tridiagonal_descr* descr);
-    void device_tridiagonal_analysis(int                  m,
-                                     int                  n,
-                                     const vector<float>& lower_diag,
-                                     const vector<float>& main_diag,
-                                     const vector<float>& upper_diag,
-                                     tridiagonal_descr*   descr);
+    void device_tridiagonal_analysis(int                   m,
+                                     int                   n,
+                                     const vector<double>& lower_diag,
+                                     const vector<double>& main_diag,
+                                     const vector<double>& upper_diag,
+                                     tridiagonal_descr*    descr);
     void device_tridiagonal_solver(int                      m,
                                    int                      n,
-                                   const vector<float>&     lower_diag,
-                                   const vector<float>&     main_diag,
-                                   const vector<float>&     upper_diag,
-                                   const vector<float>&     b,
-                                   vector<float>&           x,
+                                   const vector<double>&    lower_diag,
+                                   const vector<double>&    main_diag,
+                                   const vector<double>&    upper_diag,
+                                   const vector<double>&    b,
+                                   vector<double>&          x,
                                    const tridiagonal_descr* descr);
 }
